+++ shared/src/main/scala/pomodoro/model/PomodoroState.scala
@@ -2,5 +2,5 @@ package pomodoro.model
 
 sealed trait PomodoroState
 case object Idle extends PomodoroState
-case class Break(kind: String, started: Long) extends PomodoroState
-case class Running(started: Long) extends PomodoroState
+case class Break(kind: String, started: Double) extends PomodoroState
+case class Running(started: Double) extends PomodoroState
